Further improvement of postal code tests.
